requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
streamlit
requests